requirements.txt
asciitable==0.8.0
autopep8==1.2.4
bokeh==0.12.3
click==6.7
cycler==0.10.0
Django==1.10.4
dominate==2.3.1
Flask==0.12
Flask-Auth==0.85
Flask-Bootstrap==3.3.7.0
Flask-Login==0.4.0
Flask-WTF==0.14
gensim==1.0.0
h5py==2.6.0
image==1.5.5
itsdangerous==0.24
Jinja2==2.9.3
MarkupSafe==0.23
matplotlib==1.5.3
numpy==1.11.2
olefile==0.43
pep8==1.7.0
Pillow==4.0.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
requests==2.12.1
scipy==0.18.1
six==1.10.0
sklearn
tornado==4.4.2
tensorflow
visitor==0.1.3
vse==0.1.3
Werkzeug==0.11.15
WTForms==2.1
